add getAllPizzas() to home-routes api
