+++ server.js
@@ -3,7 +3,7 @@ const router = require('express').Router();
 const exphbs = require('express-handlebars');
 const sequelize = require("./config/connection");
 const path = require('path');
-const routes = require('./routes')
+const controllers = require('./controllers')
 
 const PORT = process.env.PORT || 3001;
 
@@ -18,7 +18,7 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use(routes);
+app.use(controllers);
 
 sequelize.sync({ force: false }).then(() => {
     app.listen(PORT, () => console.log('Now listening'));
